+++ R/auto_reprex.R
@@ -1,0 +1,8 @@
+#' @export
+auto_reprex <- function() {
+    reprex::reprex(
+      input = paste0(
+        rstudioapi::getSourceEditorContext() |>
+          purrr::pluck("contents"), "\n")
+    )
+}